requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djoser==2.0.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
kombu==4.6.11
oauthlib==3.1.0
psycopg2-binary==2.8.5
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
